refactor: restructure astro-decap-cms integration by separating type definitions and vite plugin
