+++ src/integrations/astro-decap-cms/module.d.ts
@@ -1,5 +1,5 @@
 declare module 'astro-decap-cms:config' {
-  const config: import('./index').Config
+  const config: import('./type').Config
   export = config
 }
